Add Docker, compose, and setup scripts.
This sets up docker-compose as the primary entry point to the application. All
of the dependencies (Python, Node, Postgres, libraries) are packaged up
leaving only a Docker dependency.

The docker-compose setup is split into services (which should be ran via
docker-compose up) and commands (which should be ran via docker-compose run).

Services:
* prod-mode: running the app with gunicorn and configured for production (See
  devops/prod.sh to see the startup commands)
* dev-mode: running the app with manage.py runserver, configured for
  development (See devops/dev.sh to see the startup commands). Dev mode also
  includes additional development-oriented Python libraries (See
  devops/dev_mode_install.sh).

Commands:
* manage.py
* py.test
* flake8
* pip-compile
* npm
* webpack

The Dockerfile defines a _full_ application, but when using docker-compose,
the files are overwritten with whatever's in the current directory.