Add Docker, compose, and Make -files.
The Makefile can be the primary entry point to the app (for both triggering
docker runs and for executing logic _within_ a docker/cloud.gov container).

We'll have several docker-compose files to act as configuration for three
scenarios:
1) prod-mode: running gunicorn and configured for production
2) debug-mode: running manage.py and configured for local development
3) integration-tests: like debug-mode but without a persistent volume for the
database
There's also a separate set of configs for creating a small pip-tools image.

The Dockerfile defines a _full_ application (including all of the source
code). When using docker-compose, the Python files are overwritten with
whatever's in the current directory.