Use API_URL from the environment in the JS app.
Set the API_URL on application init (either within the Node process or passed
as a `window` variable to the JS). Setting a global var like this is a bit
awkward -- if we need to set many more, we should consider adding Redux for
state management (React-redux passes the redux data store automatically).