Add minimal policy->req interface.
Copy-paste then tweak the Keywords UI into a Policies UI, where the user can
see a list of all policies, and click through to the requirements contained in
that policy. There's a lot of overlap with the Keywords code, but I figure
we'll be replacing both soon, so I'm not too worried.

Should resolve #87.