Add a lookup search/redirect endpoint.
We'd like to account for users searching for a Keyword/Policy/etc. by name
without needing client-side Javascript. The solution is to provide a
to-be-completed form which submits to a lookup page. This page will hit the
API and look for an exact string match; if it exists, the user will be
redirected to the previous page with that entry's id added to an appropriate
query parameter. If there is no direct match, we fall back to an icontains
search, which will bring up a list of entries. The user then needs to click
one to follow the same redirect process.