Add Django Debug Toolbar to development mode.
This also sets up the BrowsableAPIRenderer (which generates an HTML
version of the API results) so that there's some markup for the toolbar
to be insert into. We can still fetch JSON results by slapping on a  `.json`
extension.

This hacks a bit around a Django Debug Toolbar IP limitation (see
comments).

This requires a `docker-compose build` as it's changing the requirements