Replace custom get_queryset with custom filter.
We had been sorting by running the user-provided `sort` field through
the ORM and catching any explosion. That didn't account for paging,
unfortunately, which resulted in very slow queries. This replaces that
logic by extending the build-in OrderingFilter to include the extra
logic around finding the best match. We then specify the list of
order-able fields from our list of filters.

Two API changes:
* The query param name for sorting has changed from `sort` to `ordering`
* Bad fields are now silently dropped (I'm not a fan, but I think taking
the OrderingFilter functionality is worth this loss)