Support MAX logins.
If an appropriate environment variable is set, we replace the normal Django
login with a MAX redirect. Upon successful login, we set some of the user's
attributes.

Perhaps most interesting: we do _not_ create users during the MAX login. This
means we'll require users be created (and permissions) in the Django admin.
The alternative would be to base our permission system off which MAX
organizations a user is in (the approach DATA.gov uses). This seems like
overkill for our particular tool, however, when Django provides that
functionality out of the box.