Replace "globals" with a single "config"
Unfortunately, we can't use something like convict due to needing to
isomorphic support. This deletes the state settings in "globals" and replaces
that with a config which is derived from the Node environment (server-side) or
via window.APP_CONFIG browser-side (which is generated in the markup).

Most of the changes here are to account for a different calling scheme.