Add NewRelic browser monitoring to UI app.
The Python client does this automatically, but the Node client does not.
Ironically, we probably want statistics around the Node app's browser more
than anything. Activate it here (it degrades to an empty string when newrelic
is not enabled).

Unfortunately, the new relic code doesn't play _super_ nicely with React, but
it's much easier than needing to retrieve a public app id/key ourselves, so
this takes a shortcut and trims the string.

This also moves the config-setting `scriptTag` logic into a separate function.