Connect API to minio network.
We'd like to use a `localhost` URL when presenting uploaded files to end users
(so they can access those files outside of Docker). To do this, we'll attach
the APIs to the minio network stack. This conflicts with exposing ports
directly, however, so we need *minio* to expose those ports.