Create a base python, add external links to minio.
Combining config inheritance with the new network type caused a few commands
to blow up. Here we define a base "python" command, which is inherited by all
of the python-oriented commands/services. We also move the `external_links`
from py.test to minio as the network is being defined in minio. We use these
external links in our integration tests; it may make sense to move these
configs.