Create a "proxy" container with port-forwards.
Upon reflection, the minio-based network was useful, but was very confusing.
This replaces it with a separate container called "proxy", which defines a
shared network stack between the apis, uis, and minio (i.e. all of the things
the Docker user needs access to via `localhost`).

Unfortunately, there was some conflict on the 8000 port, so I switched `dev`
to using 8002 and mirrored that with `prod`.

Integration tests aren't working with these changes. I'll follow up.