Add a loading-screen-capable version of `resolve`.
The `resolve` decorator creates a `Resolver` element, which doesn't quite have
the hooks necessary to create a loading spinner. However, we can extend that
class to get where we need, rendering a Spinner whenever the AJAX data has not
finished.

Resolver's interface is a bit funky, so this needs to generate a handful of
functions which return functions which return functions.