Add a ConditionalRender component.
The Autocompleter code knew how to render one version (a text-box in a form)
when rendering on the server and how to convert that into a dynamic version
(using React-Select) when in a JS-capable browser.

This pulls that logic into a separate component, which will allow us to use it
more generally. It also prevents a flash of non-dynamic content by checking
whether we're rending in the browser or not initially (rather than waiting for
componentDidMount).