Adjust Topic fields, remove reqs connection.
As a first step towards extricating django-taggit, this removes model
inheritance from our Topic models. In the process, it deletes the existing
`slug` field (which we never use).

As we can't migrate between our custom TopicConnect models and Django's
many-to-many system automatically, we also remove the `topics` field from
requirements. We'll next add a many-to-many, migrate data, and delete
TopicConnects.