Migrate and delete TopicConnect data.
This is the second half of the data migration for removing taggit. Previously,
we swapped out taggit models for our own Topics; this time, we're swapping out
the custom connection between requirements and topics with the Django
many-to-many field.

We then migrate data from the custom model to the many-to-many relation and
delete the custom model.