Remove django-autocomplete-light.
We added this library to handle some of our tagging interface needs. Now that
we've switched away from taggit, this lib is no longer needed.

Note that we can't remove django-taggit until we remove references to it in
our migrations; we'll probably want to squash the migrations at some point,
which would do the trick.