Allow migrations to specify their revised models.
We now have a method to mass-create revisions for particular models via a
management command. Unfortunately, that requires we manually run the command
whenever we migrate data, which isn't super sustainable. This attempts to
resolve the issue by allowing migrations to specify the models that will need
new revisions. After migrations finish, a hook runs which inspects each
migration for this meta data and creates a revision for each.