Extend VersionAdmin with schema mismatch warnings.
When users save models, a copy of all the model's fields is stored in a
version history. As the database schema changes, that version becomes less
compatible. We can warn the user when this occurs, though, by comparing the
fields that would serialize now to the fields stored in the database.

There's a nasty snag where the serialized field names don't quite match up
when comparing the names on an unsaved Django model. This works around that by
grabbing a model from the database and showing a slightly different warning
when that's not possible.