Add a squash commit to migrations.
As a step towards removing taggit completely, we can begin to process of
squashing our migrations (which reference taggit). We have relatively
complicated migrations thus far to accommodate migrating data between multiple
schemas. Luckily, most of those don't matter with a fresh database, so our
squashed migrations is _largely_ a straight schema dump.

The two exceptions here are:
1. The "all agencies" view/fields are created by hand (Django doesn't know how
to create views)
2. There's a gin index on the search field that's similarly created by hand
(for the same reason).

The next step is to roll this migration out to everyone. Once everyone has
this migration (and runs `migrate`) we can remove migrations 0001 through 0040
(and taggit).