Prevent multiple migrations in prod.
We run multiple `api` servers in production, but we only want one of them to
be responsible for migrating our database (lest one instance's migration will
prevent the other from starting). This changeset adds a management command
which migrates the database only on the first instance when on cloud.gov, and
every time when not (e.g. local development).