Move top-level ui-related files into ui dir.
This begins the process of delineating the two apps (a Django API and a React
UI) more thoroughly by moving all of the Node-specific files into the `ui`
directory. We'll do the same with Python files shortly.

This'll provide a few benefits:
* Clearer delineation about which parts of code are used for which app.
* Our cloud.gov deploys will be smaller as we can only send up the
  app-specific code.
* Our code layout will match our system diagram.

It also paves the way for Next.js, which looks for a "pages" directory in the
*running* context. By running in the ui directory, we can use "ui/pages".