Update integration tests with the new layout.
Previously, we had been relying on a set of requirements files (and venv) in
the root of the project. We can't rely on that for integration tests now that
we've moved the Python requirements and venv into the `api` directory. The
upshot is that we can separate out the integration-test dependencies from the
dependencies for the larger app.