Remove node security platform from codeclimate.
This recently-added linter will only work if there's a shrinkwrap file in the
project root. As we've moved all the Node stuff into `ui`, that's no longer a
safe assumption.

We can add this back by running the linter outside of codeclimate, but I
figured this wasn't important since it was recently added, and we have
Gemnasium.