Add flag to Html component to prevent dynamic UI.
We'd like to reuse this Html component (which includes our stylesheets, header
nav, footer, etc.) when serving the 404 page. Unfortunately, doing so would
cause a blank screen for JS users, as React-router would get triggered, not
find the path in its routes, and explode.

To avoid this, we add a flag to prevent inclusion of browser.js. This kind of
hackery may be avoided if we switch to Next.js