Pull "new policies" API calls to top container.
As a step towards Next.js, we're trying to pull all of the asynchronous data
fetching to the top-most component (in this case, the home page). Previously,
the logic around fetching the "new policies" block had lived in the
"new-policies" container, which was one level down.

This moves that logic into a "queries.js" file so that the homepage template
can be _relatively_ logicless.