Loosen consistency guarantees in Docker.
Docker for Mac can be very slow when performing file access on a shared
directory due to consistency guarantees. This tweaks our mounted directories
so that there can be a slight delay between synchronization as the benefit of
much faster reads.

This appears to be a solution to #258 in my experience, dropping local admin
load time dropping to roughly what we see in production.

We had tried this solution before, but support wasn't fully baked.