Swap h4 for h2 in the footer.
Semantically, an h4 is "inside" an h3's content, which isn't what's actually
going on here. Instead, we should consider the footer's header as being a
direct descendent of our h1, the site's title.

This also resolves an accessibility warning by not skipping heading levels (we
didn't have an h3 on every page).