Add initial "document" app/models.
We'll try representing our documents by encoding them in the nested set model,
which required tracking a "left" and "right" attribute for each. We add a
"depth" for ease of use.

For the time being, we'll tie this to policies, though that may change if we
begin importing other types of documents.