Add a wrapper around DocNodes and trees.
Our DocNodes model doesn't keep track of its surrounding tree. While libraries
like treebeard and mptt would give us querysets for accessing siblings,
parents, etc. we don't want to hit the frequently. Instead, we'll keep the
tree in memory, encoded in a networkx graph. This'll allow us to find values
within the tree without hitting the database.

As we generally view the data one node at a time, we'll use a "cursor" pattern
to point to a specific point in the tree. We can easily navigate around the
tree using these cursors. Here, we're focusing on the methods that'll make
adding nodes to the tree easier.