Add test:watch command for front end in package.json
To make it easier to do testing, I added a test:watch command to
packge.json so that the tests will run whenever a file changes.

I had to change some of the jest configuration to get it to run in the
docker container but I think I replicated all of the functionality by
moving it into CLI arguments.