Add a Policy detail page.
The Policy detail page contains pages results of requirements for a given
policy. We want to allow users to click on particular requirements to see
their meta data. We also want to give each requirement a url so they can be
shared.

This is a partial implementation of these goals. In particular, it's missing
appropriate styling, a complete set of meta data, and a way to jump to a
particular requirement without knowing its page number in the results.
However, it lays a foundation by creating a route which lists requirements,
allowing the user to page through them, and by allowing the requirement text
to be clicked to generate a url.

We don't use a next-routes Link here because we don't want to trigger any
AJAX-like spinners.