Create <Link /> component
One link component to use everywhere so we don't need
`external-link.js` or to import from the router within our components.
This enables us to move the router around and only have to change code
in a single place.

If the link includes a protocol then it is assumed to be an external
link. The link component also ensures that the routes prop (if passed)
is a valid, registered route to prevent links breaking because of typos.

Pass through aria-* attributes to the underlying link