Propagate policy detail 404 to user.
As we've not had "detail" pages until now, we've not really needed to worry
about 404s. The policy detail is our first foray, so it's not clear we want to
handle _all_ backend 404s as frontend 404s, but we know we'd like to do this
for at least the policy fetch.