Rework Search component to be route-based.
The Search component was written prior to our Next.js-based routes, which
meant that it was doing all of its logic via substring searches. This
refactors it to instead define a constant for the default route and a mapping
between routes and their parameter names.

In the process, it alters the logic to only keep the query params when the
user is on one of the two whitelisted pages (requirements or policies); query
parameters on other pages aren't relevant.