Add Annotations abc, Footnote Citation model.
We need to reference chunks of text within a DocNode to highlight as
definitions, citations, etc. This creates an abstract base class, Annotations,
which specifies which node it's referring to and the start/end indexes in that
node's text. We then create a FootnoteCitation class as the first example of
this. These refer to footnote node.