Include policy with root document node.
The document structure doesn't have all of the meta data associated with a
policy, so we'll need to pull some in from the API. This makes that possible
by adding policy information to the document structure.

I suspect we'll want to pull the "requirement" and "policy" fields into a
separate "meta" field in the future, just to centralize that sort of
additional info away from the document structure.