Add metadata nodes to the top of Policy display.
This combines data from the Policy object and the document itself to render
the policy number, title, link, and issuance. Per @tadhg-ohiggins' suggestion,
it first searches the document nodes for this data, then falls back to the
meta data.

I'm a bit torn on doing this sort of analysis in the UI. I think, ideally, the
API would be deriving its data from the document (maybe at save time?) rather
than having the potential for conflict. Then we could treat it all as true
metadata rather than doing pseudo-parsing within the UI.