Add a field to represent node markers.
In eRegs, we included the marker as part of the text and then created an
inline layer to pull it out. Rendering this proved to be a big pain: we needed
to do lots of CSS tricks to float the marker correctly; I doubt we accounted
for all the corner cases.

As an alternative, here we follow the AKN lead (though they call it "num") and
pull the marker out into a separate field, which we can display/transform/hide
as desired.