Add table footer with footnotes.
The API returns a list of all the footnotes within a particular table, which
we can render at the bottom of the table in a `tfoot`. We need to tell it how
many columns to span, unfortunately, which means we need to do a little
calculation (here, basing it on the number of elements in the first row of the
header. We could instead use some high value, like "1000", but I'm not sure
that works on all browsers.