Add redux to app, wrapping all pages.
We'll try using redux as our state management, starting with active footnote
citations. To get there, we need to not only add the relevant libraries to our
package list, we also need to wrap each page in the `withRedux` function.
Luckily, we already have a single entrypoint in the app-wrapper.