Ensure only one footnote is opened at a time.
This adds global state to represent the open footnote. When a footnote
citation is clicked, it sets that state. When we're deciding whether to render
an inline footnote, we inspect that global state.

While splitting the store out into modules for actions, reducer,
initialization, etc. is overkill for this single bit of state, the layout is
recommended, so we sketch out its skeleton here.

This also disables the `connect` eslint rule as `connect` is frequently used
by react-redux.