Refactor annotation importer from static methods.
Previously, in determining whether a given inline tag had "special" processing
(e.g. a footnote citation), we used `getattr` on a singleton class to check if
it had a corresponding method name. This is somewhat dangerous, as classes
have many attributed in addition to defined methods. To resolve that issue,
this replaces the approach with a whitelist of methods.