Add logic for nested inlines.
Previously, we were linearlizing inline annotations so a given chunk of text
would only have one annotation applied. This extends that logic to allow
nesting of annotations in a manner similar to the DocCursor (a wrapper which
has access to the parent and child annotations).

We surface nested inline elements in a new "inlines" field while continuing to
have the "text" field refer to the whole text within an annotation. We ensure
that _all_ text is wrapped with a PlainText node at the leaves.