Propagate 404s from the API for paging errors.
Django Rest Framework returns a 404 when the user requests a non-existent page
number in a listing. Previously, we didn't account for that in the UI and just
raised the exception. This resolves by reusing our existing propagate404
function on the listing pages.

It might make sense to apply this to _all_ listing endpoints, but I've not
thought through the ramifications enough to be confident that that's a good
idea.