Add "title" field to DocNodes.
To build up a table of contents, we'd like to tag nodes within the document
tree as having "titles". This differs from specific "heading" nodes in that
the latter must be visible in the document while the former are a level of
abstraction. We discovered when working on eRegs that having a separate field
to represent a node's "title" separately from what's displayed in the document
can be very useful for navigation, search results, and other "summaries".