Refactor navigation to use redux.
Previously, we had included all of the logic around determining how many
layers of the table of contents to display within the navigation components.
We'd like to be able to access that information elsewhere, most notably to
highlight the user's current section. This changeset refactors those
components to perform the logic in a redux action creator, and to keep the
table of contents in our redux store.