Track visibleSections, currentSection in redux.
As a first step towards tracking a user's position within the table of
contents, we set up events which will trigger as the user scrolls. Whenever a
user enters or exits a section, their "currentSection" will re-calculate. This
also adds the highlighting style (currently used for hover) to ToC links
marked "active".