Adjust for floating-nav covering headings.
When on a small screen, the floating, in-document nav covers some of our text.
This causes problems when linking to specific sections (e.g. through the table
of contents). By adding a pseudo element, we can account for that.

This only adds that to sections at the moment, but we would want to do the
same for everything that's an entrypoint for users. We might create a more
generic solution then.