Always clear out missing citations.
There are times when we don't ever hit a list of footnotes but we *do* have
citations. We need to ensure those citations aren't saved (as they aren't
referencing anything).

This also switches the check from calling .footnote_node directly, as that'll
raise an exception when there is no associated doc node.