Add initial "mark" support to editor import.
This handles nested content by keeping a stack of derived markers and
linearizing them into an array of `text`s (each with the stack of marks).

Similar to the unimplemented_node, we want to track all of the meta
information around an unknown content/annotation/inline/mark type (yay
different names for similar concepts!).