Check the workflow phase rather than docnodes.
Rather than checking the _presence_ of docnodes as an indication that we
should link, we'll check the workflow phase. It continues to use a boolean
(rather than testing the workflow_phase directly in the UI) as I'm not
confident which is a better workflow. We may want to only display policies
which have an associated document going forward.

This doesn't resolve the public vs workflow question.