Initial schema, serialization for lists.
Our document model currently separates list item markers from their content --
we don't standardize to always using "1." over "(1)", etc. To mirror that in
the editor, through this schema, which has the marker be a separate chunk of
text from the list body.

As the list schema's a little complicated and largely self contained, it's
been pulled out into a separate variable.