Allow non-public documents to be edited by admins.
Previously, we weren't allowing documents associated with non-public policies
from being loaded into the editor (nor could they be retrieved in the API).
This should allow *logged in* users to see both.

It also adds a permissions check to the document load logic, which it looks
like we forgot. We don't really worry about different user groups anywhere
(other than logged in vs non), but it's best to lay that groundwork now.