Remove joinBackwards/forwards from backspace/del.
These commands would connect text elements across elements, which might not
always make sense. For example, when a user presses "delete" at the end of a
paragraph with a footnote, do we expect the footnote text to be merged with
the paragraph? If we backspace at the front of a paragraph after a heading, do
we expect the paragraph text to be placed in the heading?

Rather than figuring all of these scenarios out, seems like we should punt for
now.