Add function to derive marker from first in list.
We don't quite know what format our ordered lists will take. This accommodates
a wide variety of initial markers by looking for the last "1", "a", "i", etc.
in the first list item and treating that as a hole we can fill when
renumbering lists.

This modifies our factory methods slightly to also allow for ProseMirror
"Fragments", which are very similar to arrays.