Rework List schema to store marker info.
Previously, we had been looking at the first element in a list to determine
what markers to use. Unfortunately, that doesn't work if we're _deleting_ the
first li in a list (as the replacement marker won't contain a 1, i, a, etc.).

This takes a different approach, tracking the marker template with the
containing list.