Refactor renumberList to keep selection.
As we're modifying the nodes underneath the user's cursor, their previous
selection won't exist anymore. Luckily, the same offsets will still exist, so
we just need to set the selection to them.

This also refactors for readability: rather than needing to navigate up to the
closest list, we'll assume that we're already in a list. This makes variable
names quite a bit simpler, and doesn't affect calling too much.