Add two initial document migrations.
The PDF parser will sometimes generate documents which aren't valid in our
ProseMirror schema. This adds a command to migrate those documents to a format
which _is_ valid.

Ideally, this would include modifications to the PDF parser to prevent these
documents from being created, but we're pressed for time.

Also, we're able to cut some corners by ignoring potential changes to inline
annotations. As we're only modifying existing docnodes (or adding new ones),
we'll never break an FK. That won't be true of future doc migrations.