Add form to import PDFs, update policies.
This form takes the provided Policy and uploaded file and attempts to parse
the pdf to be associated with the Policy. We're starting with a *synchronous*
process for this import, meaning that users may receive mysterious messages if
the pdf takes too long to process. We attempt to set the Policy's workflow
phase while processing.

Along the way, we also need to run document migrations (to account for fixups
that will eventually make their way into the pdf parser), and create
Requirements. The latter is a hack to accommodate our current search, which is
based on Requirements text rather than Policy text.